backend/requirements.txt
Flask
Flask-Cors
docker
jq
kubernetes
python-dotenv
openai
pydub
pyyaml
requests
# --- Add RAG Dependencies ---
langchain
langchain-ollama
langchain-openai
langchain-community # For loaders/splitters if needed beyond core
#langchain-ibm # for Granite
faiss-cpu # Or faiss-gpu if you have CUDA setup
tiktoken # Required by langchain for token counting
# Optional: Add loaders for specific file types if needed
# pypdf
# unstructured
